refactor(prerender): extrai a carga do bulk de estados para módulo próprio
`carregarMapa`/`obterMapa` eram privados do prerender-estado.interceptor, mas
o mesmo bulk é o que a página /estados precisa. Move para
prerender-estados-bulk.ts, expondo a lista e o mapa derivado dela, ambos
memoizados — uma leitura de disco por build para os dois consumidores.

Sem mudança de comportamento.
